At some styling for closed goals
